requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
janome>=0.5.0
google-generativeai>=0.8.0
google-genai
python-dotenv>=1.0.0
